Changes to file structures and json additions -move to json folder in case of future additions -add option to see top n accounts by balance
